requirements.txt
rapidfuzz==2.13.7
requests
python-dotenv